+++ src/Eleicoes.Application/Models/PartidoViewModel.cs
@@ -2,7 +2,7 @@ namespace Eleicoes.Application.Models
 {
     public class PartidoViewModel
     {
-        public int Numero { get; set; }
+        public long Numero { get; set; }
         public string Sigla { get; set; }
         public string Nome { get; set; }
     }
